+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/OverviewRegion.java
@@ -95,7 +95,7 @@ public class OverviewRegion implements Serializable
     private int thirdParty;
 
     @Column ( name = "nps_score")
-    private float npsScore;
+    private Double npsScore;
 
     @Column ( name = "nps_detractors")
     private int npsDetractors;
@@ -374,12 +374,12 @@ public class OverviewRegion implements Serializable
     
     
 
-    public float getNpsScore()
+    public Double getNpsScore()
     {
         return npsScore;
     }
 
-    public void setNpsScore( float npsScore )
+    public void setNpsScore( Double npsScore )
     {
         this.npsScore = npsScore;
     }
